mcp-azure-devops-prod/requirements.txt
# MCP framework
mcp>=1.0.0

# Azure SDKs
azure-identity>=1.15.0
azure-mgmt-containerservice>=30.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-costmanagement>=4.0.0
azure-keyvault-secrets>=4.8.0

# Kubernetes Python client
kubernetes>=29.0.0

# HTTP client
httpx>=0.27.0

# Web server (for HTTP/SSE transport)
uvicorn>=0.29.0
starlette>=0.37.0
sse-starlette>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
